refactor(lifecycle): extract lifecycle state helpers
